Add unsigned Windows nightly releases
Merge the reviewed unsigned Windows release path and protected publisher preflight gate.

+++ src/vst3/processor.rs
@@ -706,9 +706,9 @@ unsafe fn silence_valid_stereo_output(data: &ProcessData) -> tresult {
 
 impl IProcessContextRequirementsTrait for PumpVst3Processor {
     unsafe fn getProcessContextRequirements(&self) -> u32 {
-        IProcessContextRequirements_::Flags_::kNeedTempo
-            | IProcessContextRequirements_::Flags_::kNeedProjectTimeMusic
-            | IProcessContextRequirements_::Flags_::kNeedTransportState
+        (IProcessContextRequirements_::Flags_::kNeedTempo as u32)
+            | (IProcessContextRequirements_::Flags_::kNeedProjectTimeMusic as u32)
+            | (IProcessContextRequirements_::Flags_::kNeedTransportState as u32)
     }
 }
